requirements.txt
flask
python-dotenv
supabase==2.5.0
fpdf
flask-mail
flask-session

psycopg2-binary
